batch new host messages to avoid updating the UI too often
