+++ src/gui/types/message.rs
@@ -19,8 +19,8 @@ use crate::{ChartType, IpVersion, Language, Protocol, ReportSortType, StyleType}
 pub enum Message {
     /// Every 1 second on initial page
     TickInit,
-    /// Sent by the backend parsing packets; includes the capture id and new data
-    TickRun(usize, InfoTrafficMessage),
+    /// Sent by the backend parsing packets; includes the capture id, new data, and new hosts batched data
+    TickRun(usize, InfoTrafficMessage, Vec<HostMessage>),
     /// Select adapter
     AdapterSelection(String),
     /// Select IP filter
@@ -131,6 +131,4 @@ pub enum Message {
     SetNewerReleaseStatus(Option<bool>),
     /// Set the pcap import path
     SetPcapImport(String),
-    /// Sent by the backend parsing packets; includes the capture id and new host data
-    NewHost(usize, HostMessage),
 }
